@@ -24,10 +24,10 @@ license = "GPLv2"
 readme = "README.md"
 
 [tool.poetry.dependencies]
-python = "^3.8"
+python = "^3.11"
 boto3 = "1.26.114"
 celery = "5.2.2"
-ddtrace = "1.2.1"
+ddtrace = "1.9.7"
 django-configurations = "2.3.2"
 django-debug-toolbar = "3.2.4"
 django-environ = "0.9.0"
@@ -53,7 +53,7 @@ requests = "2.31.0"
 six = "1.15.0"
 setproctitle = "1.3.2"
 social-auth-app-django = "4.0.0"
-social-auth-core = "3.3.0"
+social-auth-core = "4.4.2"
 uritemplate = "3.0.1"
 whitenoise = "5.2.0"
 
