@@ -1,6 +1,7 @@
 import csv
 import io
 import json
+from typing import Any, Optional
 
 import yaml
 from django.conf import settings
@@ -12,7 +13,7 @@ from entry.models import Entry
 from job.models import Job
 
 
-def _csv_export(job, values, recv_data, has_referral):
+def _csv_export(job: Job, values, recv_data: dict, has_referral: bool) -> Optional[io.StringIO]:
     output = io.StringIO(newline="")
     writer = csv.writer(output)
 
@@ -29,7 +30,7 @@ def _csv_export(job, values, recv_data, has_referral):
 
         # Abort processing when job is canceled
         if index % Job.STATUS_CHECK_FREQUENCY == 0 and job.is_canceled():
-            return
+            return None
 
         # Append the data which specifies Entity name to which target Entry belongs
         line_data.append(entry_info["entity"]["name"])
@@ -47,7 +48,7 @@ def _csv_export(job, values, recv_data, has_referral):
             if (value is not None) and ("type" in value):
                 vtype = value["type"]
 
-            vval = None
+            vval: Any = None
             if (value is not None) and ("value" in value):
                 vval = value["value"]
 
@@ -101,10 +102,10 @@ def _csv_export(job, values, recv_data, has_referral):
     return output
 
 
-def _yaml_export(job, values, recv_data, has_referral):
+def _yaml_export(job: Job, values, recv_data: dict, has_referral: bool) -> Optional[io.StringIO]:
     output = io.StringIO()
 
-    def _get_attr_value(atype, value):
+    def _get_attr_value(atype: int, value: dict):
         if atype & AttrTypeValue["array"]:
             return [_get_attr_value(atype ^ AttrTypeValue["array"], x) for x in value]
 
@@ -123,16 +124,16 @@ def _yaml_export(job, values, recv_data, has_referral):
         else:
             return value
 
-    resp_data = {}
+    resp_data: dict = {}
     for index, entry_info in enumerate(values):
-        data = {
+        data: dict = {
             "name": entry_info["entry"]["name"],
             "attrs": {},
         }
 
         # Abort processing when job is canceled
         if index % Job.STATUS_CHECK_FREQUENCY == 0 and job.is_canceled():
-            return
+            return None
 
         for attrinfo in recv_data["attrinfo"]:
             if attrinfo["name"] in entry_info["attrs"]:
@@ -175,10 +176,9 @@ def export_search_result(self, job_id):
     recv_data = json.loads(job.params)
 
     # Do not care whether the "has_referral" value is
-    has_referral = recv_data.get("has_referral", False)
-    referral_name = recv_data.get("referral_name")
-    entry_name = recv_data.get("entry_name")
-
+    has_referral: bool = recv_data.get("has_referral", False)
+    referral_name: Optional[str] = recv_data.get("referral_name")
+    entry_name: Optional[str] = recv_data.get("entry_name")
     if has_referral and referral_name is None:
         referral_name = ""
 
@@ -191,10 +191,9 @@ def export_search_result(self, job_id):
         referral_name,
     )
 
-    io_stream = None
+    io_stream: Optional[io.StringIO] = None
     if recv_data["export_style"] == "yaml":
         io_stream = _yaml_export(job, resp["ret_values"], recv_data, has_referral)
-
     elif recv_data["export_style"] == "csv":
         io_stream = _csv_export(job, resp["ret_values"], recv_data, has_referral)
 
