@@ -1,6 +1,7 @@
 export function formatDateTime(date: Date): string {
-  return `${date.getFullYear()}/${date.getMonth() + 1
-    }/${date.getDate()} ${date.getHours()}:${date.getMinutes()}:${date.getSeconds()}`;
+  return `${date.getFullYear()}/${
+    date.getMonth() + 1
+  }/${date.getDate()} ${date.getHours()}:${date.getMinutes()}:${date.getSeconds()}`;
 }
 
 export function formatDate(date: Date): string {
